PR Feedback: Move overridden modules into glsp-client
