+++ packages/client/src/features/command-palette/command-palette-module.ts
@@ -13,13 +13,14 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-import { CommandPalette, CommandPaletteActionProviderRegistry, FeatureModule, TYPES, bindAsService } from '@eclipse-glsp/sprotty';
+import { CommandPaletteActionProviderRegistry, FeatureModule, TYPES, bindAsService } from '@eclipse-glsp/sprotty';
 import '../../../css/command-palette.css';
+import { GlspCommandPalette } from './command-palette';
 import { CommandPaletteTool } from './command-palette-tool';
 import { ServerCommandPaletteActionProvider } from './server-command-palette-provider';
 
 export const commandPaletteModule = new FeatureModule(bind => {
-    bindAsService(bind, TYPES.IUIExtension, CommandPalette);
+    bindAsService(bind, TYPES.IUIExtension, GlspCommandPalette);
     bind(TYPES.ICommandPaletteActionProviderRegistry).to(CommandPaletteActionProviderRegistry).inSingletonScope();
     bindAsService(bind, TYPES.ICommandPaletteActionProvider, ServerCommandPaletteActionProvider);
     bindAsService(bind, TYPES.IDefaultTool, CommandPaletteTool);
