+++ packages/client/src/features/label-edit-ui/label-edit-ui-module.ts
@@ -13,8 +13,12 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
+import { EditLabelAction, EditLabelActionHandler, FeatureModule, TYPES, configureActionHandler } from '@eclipse-glsp/sprotty';
+import { GlspEditLabelUI } from './label-edit-ui';
 
-.ui-extension.hidden {
-    display: none;
-    opacity: 0;
-}
+export const labelEditUiModule = new FeatureModule((bind, unbind, isBound, rebind, ...rest) => {
+    const context = { bind, unbind, isBound, rebind };
+    configureActionHandler(context, EditLabelAction.KIND, EditLabelActionHandler);
+    bind(GlspEditLabelUI).toSelf().inSingletonScope();
+    bind(TYPES.IUIExtension).toService(GlspEditLabelUI);
+});
