+++ packages/glsp-sprotty/src/feature-modules.ts
@@ -15,16 +15,12 @@
  ********************************************************************************/
 /* eslint-disable no-restricted-imports */
 
-import { FeatureModule, bindOrRebind } from '@eclipse-glsp/protocol';
-import { EditLabelUI as SprottyEditLabelUI } from 'sprotty';
+import { FeatureModule } from '@eclipse-glsp/protocol';
 import sprottyDefaultModule from 'sprotty/lib/base/di.config';
 import sprottyButtonModule from 'sprotty/lib/features/button/di.config';
 import sprottyEdgeIntersectionModule from 'sprotty/lib/features/edge-intersection/di.config';
 import sprottyEdgeLayoutModule from 'sprotty/lib/features/edge-layout/di.config';
-import {
-    edgeEditModule as sprottyEdgeEditModule,
-    labelEditUiModule as sprottyLabelEditUiModule
-} from 'sprotty/lib/features/edit/di.config';
+import { edgeEditModule as sprottyEdgeEditModule } from 'sprotty/lib/features/edit/di.config';
 import sprottyExpandModule from 'sprotty/lib/features/expand/di.config';
 import sprottyFadeModule from 'sprotty/lib/features/fade/di.config';
 import sprottyMoveModule from 'sprotty/lib/features/move/di.config';
@@ -32,7 +28,6 @@ import sprottyOpenModule from 'sprotty/lib/features/open/di.config';
 import sprottyUpdateModule from 'sprotty/lib/features/update/di.config';
 import sprottyZorderModule from 'sprotty/lib/features/zorder/di.config';
 import sprottyModelSourceModule from 'sprotty/lib/model-source/di.config';
-import { EditLabelUI } from './ui-extension-override';
 
 export const buttonModule = new FeatureModule(sprottyButtonModule.registry);
 export const edgeEditModule = new FeatureModule(sprottyEdgeEditModule.registry);
@@ -45,12 +40,5 @@ export const moveModule = new FeatureModule(sprottyMoveModule.registry);
 export const openModule = new FeatureModule(sprottyOpenModule.registry);
 export const updateModule = new FeatureModule(sprottyUpdateModule.registry);
 export const zorderModule = new FeatureModule(sprottyZorderModule.registry);
-
-export const labelEditUiModule = new FeatureModule((bind, unbind, isBound, rebind, ...rest) => {
-    const context = { bind, unbind, isBound, rebind };
-    sprottyLabelEditUiModule.registry(bind, unbind, isBound, rebind, ...rest);
-    bind(EditLabelUI).toSelf().inSingletonScope();
-    bindOrRebind(context, SprottyEditLabelUI).toService(EditLabelUI);
-});
 
 export { sprottyDefaultModule };
